Andreas comments implemented
