+++ doc/source/getting-started/index.rst
@@ -9,4 +9,4 @@ In this section we present the basic aspects of the Qibocal design and provide i
     installation
     interface
     runcard
-    rbprotocols
+    rbprotocols/rbprotocols
