+++ src/qibocal/calibrations/__init__.py
@@ -8,4 +8,3 @@ from qibocal.calibrations.characterization.resonator_spectroscopy import *
 from qibocal.calibrations.characterization.t1 import *
 from qibocal.calibrations.protocols.crosstalkrb import qqperform_crosstalkrb
 from qibocal.calibrations.protocols.standardrb import qqperform_standardrb
-from qibocal.calibrations.protocols.test import *
