+++ src/qibocal/fitting/rb_methods.py
